Backend/restaurant_project/requirements.txt
Django==3.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-rest-knox==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
python-decouple==3.8
mysqlclient==2.2.4
PyMySQL==1.1.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
requests==2.31.0
pillow==10.4.0
psycopg2-binary==2.9.10
dj-database-url==2.2.0
gunicorn==20.1.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0